+++ im-service/im-service-aggregation/src/main/java/com/starsea/im/aggregation/service/WatchService.java
@@ -1,0 +1,18 @@
+package com.starsea.im.aggregation.service.impl;
+
+import com.starsea.im.aggregation.constant.DataSourceType;
+import com.starsea.im.aggregation.dto.WatchFormDto;
+import com.starsea.im.biz.annotation.DataSource;
+import com.starsea.im.biz.entity.WatchForm;
+
+/**
+ * Created by danny on 16/4/28.
+ */
+public interface WatchService {
+    @DataSource(DataSourceType.WRITE)
+    public  int addWatchForm(WatchForm watchForm);
+
+    @DataSource(DataSourceType.READ)
+    public WatchFormDto queryLastWatchFormByName(String name);
+
+}
