+++ im-service/im-service-biz/src/main/resources/config/mybatis/DiagnoseDao.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!DOCTYPE mapper PUBLIC
+        "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.starsea.im.biz.dao.DiagnoseDao">
+
+    <insert id="addStudyFormByTeacher" parameterType="StudyForm">
+        INSERT INTO test.StudyForm_Teacher(NAME, Age, Sex, Class, School, Organization, EvaluationPerson, EvaluationTime,
+        Question1, Question2, Question3, Question4, Question5, Question6,
+        Question7, Question8, Question9, Question10, Question11, Question12,
+        Question13, Question14, Question15, Question16, Question17, Question18)
+        VALUES (#{name}, #{age}, #{sex}, #{banji},
+        #{school}, #{organization}, #{evaluationPerson},
+        #{evaluationTime}, #{Question1}, #{Question2},
+        #{Question3}, #{Question4}, #{Question5}, #{Question6},
+        #{Question7}, #{Question8}, #{Question9}, #{Question10},
+        #{Question11}, #{Question12}, #{Question13}, #{Question14},
+        #{Question15}, #{Question16}, #{Question17}, #{Question18});
+    </insert>
+
+    <select id="queryStudyFormByTeacher" resultType="StudyForm">
+        SELECT * FROM test.StudyForm_Teacher
+    </select>
+
+    <select id="queryLastStudyFormByName" resultType="StudyForm">
+       SELECT * FROM test.StudyForm_Teacher where Name = #{name} order by id desc limit 1
+    </select>
+
+</mapper>
