+++ im-service/im-service-biz/src/main/resources/config/mybatis/mybatis-config.xml
@@ -5,5 +5,7 @@
 <configuration>
     <mappers>
         <mapper resource="config/mybatis/UserDao.xml"/>
+        <mapper resource="config/mybatis/DiagnoseDao.xml"/>
+        <mapper resource="config/mybatis/WatchDao.xml"/>
     </mappers>
 </configuration>
