+++ im-service/im-service-web/src/main/webapp/studyForm.html
@@ -682,7 +682,7 @@
 
 
         $.ajax({
-            url: '../starsea/HelloWeb/addForm',
+            url: '../starsea/diagnose/addForm',
             data: {
                 myName:myName,
                 sex:sex,
